Split up the module for the connection and sending/generating messages
- Set a `source_addr` for the `submit_sm` message
