+++ config/config.exs
@@ -26,8 +26,9 @@ config :voomex, Voomex.SMPP,
   start: true,
   host: "localhost",
   port: 2775,
-  system_id: "10030",
-  password: "pass30",
+  system_id: "smppclient1",
+  password: "password",
+  source_addr: "12345",
   source_ton: 1,
   source_npi: 1,
   dest_ton: 1,
